Restructured gui package
