+++ src/gui/controllers/GameController.java
@@ -1,5 +1,4 @@
-
-package gui;
+package gui.controllers;
 
 import java.net.URL;
 import java.util.Optional;
