+++ src/gui/views/Computer.fxml
@@ -11,7 +11,7 @@
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.text.Font?>
 
-<GridPane prefHeight="400.0" prefWidth="600.0" style="-fx-background-color: #FFFFFF;" xmlns="http://javafx.com/javafx/8.0.141" xmlns:fx="http://javafx.com/fxml/1" fx:controller="gui.ComputerWindowController">
+<GridPane prefHeight="400.0" prefWidth="600.0" style="-fx-background-color: #FFFFFF;" xmlns="http://javafx.com/javafx/8.0.141" xmlns:fx="http://javafx.com/fxml/1" fx:controller="gui.controllers.ComputerWindowController">
   <columnConstraints>
     <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
     <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
@@ -21,17 +21,17 @@
     <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
   </rowConstraints>
    <children>
-      <Button mnemonicParsing="false" onAction="#onTweetClicked" styleClass="window-button-computer" stylesheets="@styles.css" text="Tweet" GridPane.halignment="CENTER" GridPane.rowIndex="1" GridPane.valignment="CENTER">
+      <Button mnemonicParsing="false" onAction="#onTweetClicked" styleClass="window-button-computer" stylesheets="@../resources/styles.css" text="Tweet" GridPane.halignment="CENTER" GridPane.rowIndex="1" GridPane.valignment="CENTER">
          <GridPane.margin>
             <Insets right="100.0" top="150.0" />
          </GridPane.margin>
       </Button>
-      <Button mnemonicParsing="false" onAction="#onRandomTweetClicked" styleClass="window-button-computer" stylesheets="@styles.css" text="Random tweet" GridPane.halignment="CENTER" GridPane.rowIndex="1" GridPane.valignment="CENTER">
+      <Button mnemonicParsing="false" onAction="#onRandomTweetClicked" styleClass="window-button-computer" stylesheets="@../resources/styles.css" text="Random tweet" GridPane.halignment="CENTER" GridPane.rowIndex="1" GridPane.valignment="CENTER">
          <GridPane.margin>
             <Insets left="100.0" top="150.0" />
          </GridPane.margin>
       </Button>
-      <TextArea fx:id="writeTweetArea" prefHeight="152.0" prefWidth="323.0" style="-fx-background-color: #c0deed;" stylesheets="@styles.css" GridPane.columnSpan="2" GridPane.rowIndex="1">
+      <TextArea fx:id="writeTweetArea" prefHeight="152.0" prefWidth="323.0" style="-fx-background-color: #c0deed;" stylesheets="@../resources/styles.css" GridPane.columnSpan="2" GridPane.rowIndex="1">
          <GridPane.margin>
             <Insets bottom="50.0" left="30.0" right="30.0" top="50.0" />
          </GridPane.margin>
@@ -59,7 +59,7 @@
             <Insets right="230.0" />
          </GridPane.margin>
          <image>
-            <Image url="@resources/Twitter_Logo_Blue.png" />
+            <Image url="@../resources/Twitter_Logo_Blue.png" />
          </image>
       </ImageView>
    </children>
